Similar drill for LDM. Now abstracted UEnc classes.
