+++ src/dymad/models/__init__.py
@@ -1,6 +1,7 @@
 # Import all models
 from dymad.models.model_base import ModelBase
 from dymad.models.model_temp_ucat import ModelTempUCat, ModelTempUCatGraph
+from dymad.models.model_temp_uenc import ModelTempUEnc, ModelTempUEncGraph
 from dymad.models.kbf import DGKBF, DKBF, KBF, GKBF
 from dymad.models.km import DGKM, DGKMSK, DKM, DKMSK, GKM, KM
 from dymad.models.ldm import DGLDM, DLDM, GLDM, LDM
@@ -18,6 +19,8 @@ __all__ = [
     "ModelBase",
     "ModelTempUCat",
     "ModelTempUCatGraph",
+    "ModelTempUEnc",
+    "ModelTempUEncGraph",
     "GKBF",
     "GKM",
     "GLDM",
